+++ zend/includes.h
@@ -59,51 +59,51 @@
 /**
  *  Include other files from this library
  */
-#include "../include/visibility.h"
-#include "../include/deprecated.h"
-#include "../include/noexcept.h"
-#include "../include/thread_local.h"
-#include "../include/platform.h"
-#include "../include/version.h"
-#include "../include/inivalue.h"
-#include "../include/ini.h"
-#include "../include/exception.h"
-#include "../include/fatalerror.h"
-#include "../include/streams.h"
-#include "../include/type.h"
-#include "../include/errors.h"
-#include "../include/hashparent.h"
-#include "../include/value.h"
-#include "../include/valueiterator.h"
-#include "../include/array.h"
-#include "../include/object.h"
-#include "../include/globals.h"
-#include "../include/argument.h"
-#include "../include/byval.h"
-#include "../include/byref.h"
-#include "../include/global.h"
-#include "../include/hashmember.h"
-#include "../include/super.h"
-#include "../include/parameters.h"
-#include "../include/modifiers.h"
-#include "../include/base.h"
-#include "../include/countable.h"
-#include "../include/arrayaccess.h"
-#include "../include/serializable.h"
-#include "../include/iterator.h"
-#include "../include/traversable.h"
-#include "../include/classtype.h"
-#include "../include/classbase.h"
-#include "../include/interface.h"
-#include "../include/constant.h"
-#include "../include/zendcallable.h"
-#include "../include/class.h"
-#include "../include/namespace.h"
-#include "../include/extension.h"
-#include "../include/call.h"
-#include "../include/script.h"
-#include "../include/file.h"
-#include "../include/function.h"
+#include "../include/phpcpp/visibility.h"
+#include "../include/phpcpp/deprecated.h"
+#include "../include/phpcpp/noexcept.h"
+#include "../include/phpcpp/thread_local.h"
+#include "../include/phpcpp/platform.h"
+#include "../include/phpcpp/version.h"
+#include "../include/phpcpp/inivalue.h"
+#include "../include/phpcpp/ini.h"
+#include "../include/phpcpp/exception.h"
+#include "../include/phpcpp/fatalerror.h"
+#include "../include/phpcpp/streams.h"
+#include "../include/phpcpp/type.h"
+#include "../include/phpcpp/errors.h"
+#include "../include/phpcpp/hashparent.h"
+#include "../include/phpcpp/value.h"
+#include "../include/phpcpp/valueiterator.h"
+#include "../include/phpcpp/array.h"
+#include "../include/phpcpp/object.h"
+#include "../include/phpcpp/globals.h"
+#include "../include/phpcpp/argument.h"
+#include "../include/phpcpp/byval.h"
+#include "../include/phpcpp/byref.h"
+#include "../include/phpcpp/global.h"
+#include "../include/phpcpp/hashmember.h"
+#include "../include/phpcpp/super.h"
+#include "../include/phpcpp/parameters.h"
+#include "../include/phpcpp/modifiers.h"
+#include "../include/phpcpp/base.h"
+#include "../include/phpcpp/countable.h"
+#include "../include/phpcpp/arrayaccess.h"
+#include "../include/phpcpp/serializable.h"
+#include "../include/phpcpp/iterator.h"
+#include "../include/phpcpp/traversable.h"
+#include "../include/phpcpp/classtype.h"
+#include "../include/phpcpp/classbase.h"
+#include "../include/phpcpp/interface.h"
+#include "../include/phpcpp/constant.h"
+#include "../include/phpcpp/zendcallable.h"
+#include "../include/phpcpp/class.h"
+#include "../include/phpcpp/namespace.h"
+#include "../include/phpcpp/extension.h"
+#include "../include/phpcpp/call.h"
+#include "../include/phpcpp/script.h"
+#include "../include/phpcpp/file.h"
+#include "../include/phpcpp/function.h"
 
 /**
  *  Common header files for internal use only
